better import strategy, use short month handler

+++ components/person-card.js
@@ -3,7 +3,7 @@ personCardStyleSheet.replaceSync(/*css*/ `
 	:host {
 		display: block;
 	}
-	
+
 	person-card-container {
 		display: flex;
 		flex-direction: column;
@@ -12,7 +12,7 @@ personCardStyleSheet.replaceSync(/*css*/ `
 		width: 100px;
 		align-items: center;
 	}
-	
+
 	person-card-name {
 		text-overflow: ellipsis;
 		white-space: nowrap;
@@ -43,7 +43,7 @@ class PersonCard extends HTMLElement {
 				<wa-avatar label="To be built" image="${this.getAttribute('avatar')}"></wa-avatar>
 				<person-card-name>${truncatedName}</person-card-name>
 				<wa-tag size="small" pill>
-					<wa-format-date month="long" day="numeric" year="numeric" date="${this.getAttribute('date')}"></wa-format-date>
+					<wa-format-date month="short" day="numeric" year="numeric" date="${this.getAttribute('date')}"></wa-format-date>
 				</wa-tag>
 			</person-card-container>
 		`;
